+++ CHANGELOG.md
@@ -1,9 +1,21 @@
+## 1.0.3 (February 10, 2022)
+
+ENHANCEMENT:
+
+  * Added input for multi-region
+  * Added input for key rotation
+  * Added input for deletion window
+  * Added input for key usage
+  * Updated README
+  * Updated examples
+
 ## 1.0.2 (February 10, 2022)
 
 ENHANCEMENT:
 
   * Added descriptions to outputs.
   * Updated README
+
 ## 1.0.1 (February 10, 2022)
 
 ENHANCEMENT: